+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [2.0.0] - 2020-06-15
 ### BREAKING
 - Refactored the IntuneBackupAndRestore module to depend on Microsoft.Graph.Intune module, instead of the custom MSGraphFunctions module.
+- Fixed an issue where restoring assignments could result in an error. Now also supports restoring assignments for Line-of-Business Client Apps.
+
 
 ## [1.5.0] - 2020-04-02
 - Added function `Invoke-IntuneBackupDeviceManagementIntent`.